requirements.txt
flask==3.0.0
requests==2.31.0
cryptography==41.0.7
python-dotenv==1.0.0
supabase==2.0.0

